Moved oShoe to be a child of oTools_Parent
I also made a new script for analysing list inputs and used it for a resolution drop down list in the oCamera variable definitions.

Fixed a bit more of the parallax

Removed the old landing in shoe check.

set some default variable definitions in oSolid

set default collision object list for all calls of physics() to be the collision_object_array of the object running the script